resolve all warnings
