+++ three_by_three/src/cube_ops/partial_reprs/corner_orient.rs
@@ -1,4 +1,4 @@
-use crate::kociemba::coords::coords::CornerOrientRawCoord;
+use crate::kociemba::coords::CornerOrientRawCoord;
 
 use super::corner_perm::CornerPerm;
