+++ three_by_three/src/cube_ops/partial_reprs/corner_perm.rs
@@ -1,6 +1,4 @@
-use crate::{
-    kociemba::coords::coords::CornerPermRawCoord, permutation_math::permutation::Permutation,
-};
+use crate::{kociemba::coords::CornerPermRawCoord, permutation_math::permutation::Permutation};
 
 /// The slot representation for corner permutation.
 /// While `Permutation<N>` represents an element of the permutation group, this represents
