+++ three_by_three/src/cube_ops/partial_reprs/edge_orient.rs
@@ -1,4 +1,4 @@
-use crate::kociemba::coords::coords::EdgeOrientRawCoord;
+use crate::kociemba::coords::EdgeOrientRawCoord;
 
 use super::edge_perm::EdgePerm;
