+++ three_by_three/src/kociemba/coords/corner_perm_combo_coord.rs
@@ -1,7 +1,7 @@
 use crate::{
     cube_ops::{cube_move::CubeMove, cube_sym::DominoSymmetry},
     kociemba::{
-        coords::coords::{CornerPermRawCoord, CornerPermSymCoord},
+        coords::{CornerPermRawCoord, CornerPermSymCoord},
         tables::{
             lookup_sym_corner_perm::LookupSymCornerPermTable,
             move_sym_corner_perm::MoveSymCornerPermTable,
