+++ three_by_three/src/kociemba/partial_reprs/edge_group_orient.rs
@@ -2,7 +2,7 @@ use crate::{
     cube_ops::{
         cube_move::CubeMove, cube_sym::DominoSymmetry, partial_reprs::edge_orient::EdgeOrient,
     },
-    kociemba::coords::coords::{EdgeGroupOrientRawCoord, EdgeGroupRawCoord, EdgeOrientRawCoord},
+    kociemba::coords::{EdgeGroupOrientRawCoord, EdgeGroupRawCoord, EdgeOrientRawCoord},
 };
 
 use super::edge_group::EdgeGroup;
