+++ three_by_three/src/kociemba/tables/move_raw_corner_orient.rs
@@ -9,7 +9,7 @@ use crate::{
     cube_ops::{
         cube_move::CubeMove, cube_sym::DominoSymmetry, partial_reprs::corner_orient::CornerOrient,
     },
-    kociemba::coords::coords::CornerOrientRawCoord,
+    kociemba::coords::CornerOrientRawCoord,
 };
 
 use super::table_loader::load_table;
