+++ three_by_three/src/kociemba/tables/move_sym_edge_group_orient.rs
@@ -8,8 +8,7 @@ use crate::{
     cube_ops::{cube_move::CubeMove, cube_sym::DominoSymmetry},
     kociemba::{
         coords::{
-            coords::EdgeGroupOrientSymCoord,
-            edge_group_orient_combo_coord::EdgeGroupOrientComboCoord,
+            EdgeGroupOrientSymCoord, edge_group_orient_combo_coord::EdgeGroupOrientComboCoord,
         },
         partial_reprs::edge_group_orient::EdgeGroupOrient,
         tables::table_loader::{as_u16_slice, as_u16_slice_mut},
