+++ three_by_three/src/kociemba/tables/prune_phase_2_corner_sym.rs
@@ -6,7 +6,7 @@ use std::path::Path;
 use anyhow::Result;
 use memmap2::Mmap;
 
-use crate::kociemba::coords::coords::{CornerPermSymCoord, UDEdgePermRawCoord};
+use crate::kociemba::coords::{CornerPermSymCoord, UDEdgePermRawCoord};
 use crate::kociemba::tables::prune_phase_2::PrunePhase2Table;
 
 use super::table_loader::load_table;
